+++ applications/fishing-map/src/features/workspace/environmental/EnvironmentalSection.tsx
@@ -6,13 +6,11 @@ import { DatasetCategory } from '@globalfishingwatch/api-types'
 import { selectEnvironmentalDataviews } from 'features/workspace/workspace.selectors'
 import styles from 'features/workspace/shared/Sections.module.css'
 import NewDatasetTooltip from 'features/datasets/NewDatasetTooltip'
-import { isGuestUser } from 'features/user/user.selectors'
 import TooltipContainer from 'features/workspace/shared/TooltipContainer'
 import LayerPanel from './EnvironmentalLayerPanel'
 
 function EnvironmentalLayerSection(): React.ReactElement | null {
   const { t } = useTranslation()
-  const guestuser = useSelector(isGuestUser)
   const [newDatasetOpen, setNewDatasetOpen] = useState(false)
   const dataviews = useSelector(selectEnvironmentalDataviews)
 
@@ -24,40 +22,28 @@ function EnvironmentalLayerSection(): React.ReactElement | null {
     <div className={styles.container}>
       <div className={styles.header}>
         <h2 className={styles.sectionTitle}>{t('common.environment', 'Environment')}</h2>
-        {guestuser ? (
+        <TooltipContainer
+          visible={newDatasetOpen}
+          onClickOutside={() => {
+            setNewDatasetOpen(false)
+          }}
+          component={
+            <NewDatasetTooltip
+              datasetCategory={DatasetCategory.Environment}
+              onSelect={() => setNewDatasetOpen(false)}
+            />
+          }
+        >
           <IconButton
             icon="plus"
             type="border"
             size="medium"
-            disabled
-            tooltip={t('dataset.uploadLogin', 'You need to log in to upload datasets')}
+            tooltip={t('dataset.addEnvironmental', 'Add environmental dataset')}
             tooltipPlacement="top"
             className="print-hidden"
+            onClick={onAddClick}
           />
-        ) : (
-          <TooltipContainer
-            visible={newDatasetOpen}
-            onClickOutside={() => {
-              setNewDatasetOpen(false)
-            }}
-            component={
-              <NewDatasetTooltip
-                datasetCategory={DatasetCategory.Environment}
-                onSelect={() => setNewDatasetOpen(false)}
-              />
-            }
-          >
-            <IconButton
-              icon="plus"
-              type="border"
-              size="medium"
-              tooltip={t('dataset.addEnvironmental', 'Add environmental dataset')}
-              tooltipPlacement="top"
-              className="print-hidden"
-              onClick={onAddClick}
-            />
-          </TooltipContainer>
-        )}
+        </TooltipContainer>
       </div>
       {dataviews?.map((dataview) => (
         <LayerPanel key={dataview.id} dataview={dataview} />