+++ applications/fishing-map/src/features/workspace/vessels/VesselLayerPanel.tsx
@@ -204,7 +204,7 @@ function LayerPanel({ dataview }: LayerPanelProps): React.ReactElement {
                       if (!fieldValue) return null
                       return (
                         <li key={field.id} className={styles.infoContentItem}>
-                          <label>{t(`vessel.${field.id}`)}</label>
+                          <label>{t(`vessel.${field.id}` as any)}</label>
                           <span>
                             {field.type === 'date' ? (
                               <I18nDate date={fieldValue} />