+++ packages/react-hooks/src/use-map-interaction/use-map-interaction.ts
@@ -50,6 +50,7 @@ const getExtendedFeatures = (
         )
         if (!values || !values.filter((v) => v > 0).length) return []
 
+        const visibleSublayers = metadata?.temporalgrid?.visibleSublayers as boolean[]
         return values.flatMap((value: any, i: number) => {
           if (value === 0) return []
           return [
@@ -57,6 +58,7 @@ const getExtendedFeatures = (
               ...extendedFeature,
               temporalgrid: {
                 sublayerIndex: i,
+                visible: visibleSublayers[i] === true,
                 col: properties._col as number,
                 row: properties._row as number,
               },